LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-invalid-http-http-expected.txt

CONSOLE MESSAGE: The 'referrer' Content Security Policy directive has the invalid value "invalid". Valid values are "no-referrer", "no-referrer-when-downgrade", "origin", "origin-when-cross-origin", and "unsafe-url".
Test that invalid is ignored from HTTP to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is "http://127.0.0.1:8000/security/contentSecurityPolicy/resources/referrer-test.php?policy=invalid&to=http"
PASS successfullyParsed is true

TEST COMPLETE
